added users Router
